+++ api/helpers/resolvers/rolename.js
@@ -50,40 +50,57 @@ module.exports = {
       default:
         return await new Promise(async (resolve, reject) => {
           var children = [];
+          var _children = [];
+          var children2 = [];
+          var childrenMain = [];
           querySearch.forEach((option) => {
-            children.push({
-              title: option.name,
-              function: (outputMessage, senderMessage) => {
-                outputMessage.delete();
-                return resolve(option);
-              },
-            });
+            children.push(option);
+            childrenMain.push(option);
           });
 
-          const menuTemplate = {
-            title: ":question: Choose a role",
-            description: `Multiple roles match the role you provided: ${inputs.roleName}. Please click the reaction corresponding to the role you want.`,
-            color: "BLUE",
-            footer: {
-              type: "timestamp",
-              value: Date.now(),
-            },
-            children: children,
-          };
-
-          var response = await inputs.message.reply(`Please wait...`);
-          const menu = new Discord.DiscordMenu(
-            menuTemplate,
-            response,
-            inputs.message,
-            {
-              dataPersistance: true,
-              backButton: true,
-              time: 60,
+          // Now, break the roles up into groups of 10 for pagination.
+          while (children.length > 0) {
+            _children.push(children.shift());
+            if (_children.length > 9) {
+              children2.push(_.cloneDeep(_children));
+              _children = [];
             }
-          );
+          }
+          if (_children.length > 0) {
+            children2.push(_.cloneDeep(_children));
+          }
 
-          menu.start();
+          const menu = new Discord.DiscordMenu(
+            inputs.message.channel,
+            inputs.message.author.id,
+            children2.map((group) => {
+              var groupEmbed = new Discord.MessageEmbed()
+                .setAuthor(
+                  `${inputs.message.author.tag}`,
+                  `${inputs.message.author.displayAvatarURL()}`
+                )
+                .setTitle(`Multiple roles found!`)
+                .setDescription(
+                  `Multiple roles matched the name **${possible.name}**. Use the menu to find which role you meant, and then type its name in a message.`
+                )
+                .setColor(`#8800FF`)
+                .setFooter(`User ID: ${inputs.message.author.id}`)
+                .setTimestamp();
+              group.map((child) => {
+                groupEmbed.addField(child.name, `ID: ${child.id}`);
+              });
+              return groupEmbed;
+            }),
+            childrenMain.map((child) => {
+              return {
+                message: child.name,
+                fn: (senderMessage) => {
+                  senderMessage.delete();
+                  return resolve(child);
+                },
+              };
+            })
+          );
         });
     }
   },