+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a list of changes made to the Drago's Moderations Bot.
 
+## 0.7.0-alpha.12 - UNRELEASED
+
+### Fixed
+
+- rolename and username resolvers.
+- Grant command often did not apply permissions the first time it was used.
+
 ## 0.7.0-alpha.11 - 2020-07-30
 
 ### Deprecated